Move prometheus pkg to a separate pkg
